Move src/__tests__/ to src/components/__tests__/
